% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.genclust.em.R
\name{AIC.genclust.em}
\alias{AIC.genclust.em}
\title{Compute Akaike Information Criterion (AIC) for genclust.em}
\usage{
\method{AIC}{genclust.em}(object, ...)
}
\arguments{
\item{object}{An object returned by the function \code{\link{genclust.em}}.}

\item{...}{Further arguments for compatibility with the \code{AIC} generic
(currently not used).}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\code{\link{genclust.em}} to generate clustering solutions.
}

