% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.snapclust.R
\name{AIC.snapclust}
\alias{AIC.snapclust}
\title{Compute Akaike Information Criterion (AIC) for snapclust}
\usage{
\method{AIC}{snapclust}(object, ...)
}
\arguments{
\item{object}{An object returned by the function \code{\link{snapclust}}.}

\item{...}{Further arguments for compatibility with the \code{AIC} generic
(currently not used).}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\seealso{
\code{\link{snapclust}} to generate clustering solutions.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
