% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.genclust.em.R
\name{BIC.genclust.em}
\alias{BIC.genclust.em}
\title{Compute Bayesian Information Criterion (BIC) for genclust.em}
\usage{
\method{BIC}{genclust.em}(object, ...)
}
\arguments{
\item{object}{An object returned by the function \code{\link{genclust.em}}.}

\item{...}{Further arguments for compatibility with the \code{BIC} generic
(currently not used).}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\seealso{
\code{\link{genclust.em}} to generate clustering solutions.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
