% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_C_routines.R
\docType{data}
\name{.internal_C_routines}
\alias{.internal_C_routines}
\alias{CheckAllSeg}
\alias{GLdotProd}
\alias{GLsumFreq}
\alias{GLsumInt}
\alias{binIntToBytes}
\alias{bytesToBinInt}
\alias{bytesToInt}
\alias{nb_shared_all}
\title{Internal C routines}
\format{An object of class \code{NULL} of length 0.}
\usage{
.internal_C_routines
}
\description{
These functions are internal C routines used in adegenet. Do not use them
unless you know what you are doing.
}
\author{
Thibaut Jombart
}
\keyword{datasets}
