% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newclust.R
\name{genclust.emmcmc}
\alias{genclust.emmcmc}
\alias{plot.emmcmc}
\title{EM-MCMC algorithm for genetic clustering}
\usage{
genclust.emmcmc(x, k, n.iter = 100, sample.every = 10, pop.ini = NULL,
  max.em.iter = 20, prop.move = 0.2, detailed)

plot.emmcmc(x, y = "ll", type = c("trace", "hist", "density", "groups"),
  burnin = 0)
}
\arguments{
\item{x}{a \linkS4class{genind} object}

\item{k}{the number of clusters to look for}

\item{n.iter}{the number of iterations of the MCMC}

\item{sample.every}{the frequency of sampling from the MCMC}

\item{pop.ini}{an optional factor defining the initial cluster configuration}

\item{max.em.iter}{the maximum number of iterations for the EM algorithm}

\item{prop.move}{the proportion of individuals moved across groups at each MCMC iteration}

\item{detailed}{a logical stating whether extra details should be incorporated into the output;
these include group membership probability, indication of convergence, and the number of
iterations used before convergence}
}
\description{
Do not use.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

