% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_webapp.R
\name{export_to_webapp}
\alias{export_to_webapp}
\alias{export_to_webapp.default}
\alias{export_to_webapp.dapc}
\alias{export_to_webapp.dudi}
\alias{export_to_webapp.spca}
\title{Export analysis for webapp visualisation}
\usage{
export_to_webapp(x, ...)

\method{export_to_webapp}{default}(x, ...)

\method{export_to_webapp}{dapc}(x, info, ...)

\method{export_to_webapp}{dudi}(x, info, ...)

\method{export_to_webapp}{spca}(x, info, ...)
}
\arguments{
\item{x}{The analysis to be exported.}

\item{...}{Further arguments to pass to other methods.}

\item{info}{A data.frame with additional information containing at least the
following columns: \code{key} (individual label), \code{latitude}, and
\code{longitude}.}
}
\description{
Export analysis for webapp visualisation
}
\examples{

data(sim2pop)

dapc1 <- dapc(sim2pop, n.pca = 10, n.da = 1)

info <- data.frame(key = indNames(sim2pop),
                   lat = other(sim2pop)$xy[,2],
                   lon = other(sim2pop)$xy[,1],
                   Population = pop(sim2pop))

out <- export_to_webapp(dapc1, info)
head(out)

data(rupica)

spca1 <- spca(rupica, type=5, d1 = 0, d2 = 2300,
              plot = FALSE, scannf = FALSE,
              nfposi = 2,nfnega = 0)

info <- data.frame(key = indNames(rupica),
                   lat = rupica$other$xy[,2],
                   lon = rupica$other$xy[,1])

out <- export_to_webapp(spca1, info)
head(out)

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
