% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genclust.em.choose.k.R
\name{genclust.em.choose.k}
\alias{genclust.em.choose.k}
\title{Choose the number of clusters for genclust.em using BIC}
\usage{
genclust.em.choose.k(max, ..., IC = AIC, IC.only = TRUE)
}
\arguments{
\item{max}{An integer indicating the maximum number of clusters to seek;
\code{\link{genclust.em}} will be run for all k from 2 to max.}

\item{...}{Arguments passed to \code{\link{genclust.em}}.}

\item{IC}{A function computing the information criterion for
\code{\link{BIC.genclust.em}} objects. Available statistics are
\code{AIC} (default) and \code{BIC}.}

\item{IC.only}{A logical (TRUE by default) indicating if IC values only
should be returned; if \code{FALSE}, full \code{genclust.em} objects are
returned.}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\seealso{
\code{\link{genclust.em}} to generate individual clustering
    solutions, and \code{\link{BIC.genclust.em}} for computing BIC for
    \code{genclust.em} objects.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
