% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoplot.R, R/newclust.R
\name{compoplot.genclust.em}
\alias{compoplot.genclust.em}
\alias{genclust.em}
\title{Maximum-likelihood genetic clustering using EM algorithm}
\usage{
\method{compoplot}{genclust.em}(x, ...)

genclust.em(x, k, pop.ini = NULL, max.iter = 100, n.start = 10,
  detailed = TRUE)
}
\arguments{
\item{x}{a \linkS4class{genind} object}

\item{k}{the number of clusters to look for}

\item{pop.ini}{an optional factor defining the initial cluster configuration}

\item{max.iter}{the maximum number of iteration of the EM algorithm}

\item{n.start}{the number of times the EM algorithm is run, each time with different random
starting conditions}

\item{detailed}{a logical stating whether extra details should be incorporated into the output;
these include group membership probability, indication of convergence, and the number of
iterations used before convergence}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\examples{
data(microbov)

## try function using k-means initialization
grp.ini <- find.clusters(microbov, n.clust=15, n.pca=150)

## run EM algo
res <- genclust.em(microbov, 15, pop.ini = grp.ini$grp, detailed = TRUE)
names(res)
res$converged
res$n.iter


}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com} and Marie-Pauline Beugin
}

