\name{haploGen}
\alias{haploGen}
\alias{print.haploGen}
\alias{[.haploGen}
\alias{labels.haploGen}
\alias{as.POSIXct.haploGen}
\alias{seqTrack.haploGen}
\alias{haploGen-class}
% \alias{seqTrackG.haploGen}
% \alias{optimize.seqTrack.haploGen}
\alias{as.seqTrack.haploGen}
\alias{plotHaploGen}
\alias{sample.haploGen}
\alias{as,haploGen,graphNEL-method}
\alias{coerce,haploGen,graphNEL-method}
\title{Simulation of populations of haplotypes}
\description{
  Important: these functions are parts of a publication currently under
  review. They will be documented once accepted for publication.
  Please email the author if you are interested in using it.
}
% \usage{

% }
% \arguments{
%   % \item{}{}
% }
\author{Thibaut Jombart \email{t.jombart@imperial.ac.uk}}
