\name{haploPop}
\alias{haploPop}
\alias{print.haploPop}
\alias{summary.haploPop}
\alias{sample.haploPop}
\alias{dist.haploPop}
\alias{plot.haploPop}
\alias{haploPopDiv}
\title{Simulation of populations of haplotypes}
\description{
  Important: these functions are parts of a publication currently under
  review. They will be documented once accepted for publication.
  Please email the author if you are interested in using it.
}
% \usage{

% }
% \arguments{
%   % \item{}{}
% }
\author{Thibaut Jombart \email{t.jombart@imperial.ac.uk}}
