\name{seqTrack}
\alias{seqTrack}
\alias{plotSeqTrack}
\title{seqTrack algorithm}
\description{
  The \code{seqTrack} algorithm is currently in development. Please do
  not use it.
}
\usage{
seqTrack(seq.names, seq.dates, W, optim = c("min", "max"), \dots)
plotSeqTrack(x, xy, useArrows = TRUE, col = NULL, bg = "grey", add = FALSE, 
    \dots)
}
\arguments{
  \item{seq.names}{}
  \item{seq.dates}{}
  \item{W}{}
  \item{optim}{}
  \item{x}{}
  \item{xy}{}
  \item{useArrows}{}
  \item{col}{}
  \item{bg}{}
  \item{add}{}
  \item{\dots}{}
  % \item{}{}
}
\author{Thibaut Jombart \email{t.jombart@imperial.ac.uk}}
