% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{virtualClasses}
\alias{callOrNULL-class}
\alias{charOrNULL-class}
\alias{factorOrNULL-class}
\alias{gen-class}
\alias{indInfo-class}
\alias{intOrNULL-class}
\alias{intOrNum-class}
\alias{listOrNULL-class}
\alias{popInfo-class}
\alias{virtualClasses}
\title{Virtual classes for adegenet}
\description{
These virtual classes are only for internal use in adegenet
}
\section{Objects from the Class}{
 A virtual Class: No objects may be created
from it.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}
\keyword{classes}

