% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxil.R
\docType{methods}
\name{Website and tutorials}
\alias{Website and tutorials}
\alias{adegenetTutorial}
\alias{adegenetTutorials}
\alias{adegenetWeb}
\title{Website and tutorials for adegenet}
\usage{
adegenetWeb()
}
\arguments{
\item{which}{a character string indicating the type of tutorial to open}
}
\description{
These functions allow to access documentation for adegenet available
online.\cr
}
\details{
These functions include: \itemize{ \item \code{adegenetWeb}: opens the
adegenet website in a web navigator.  \item \code{adegenetTutorial}: opens a
specified tutorial.  }
}
\examples{
\dontrun{
## this opens the adegenet website
adegenetWeb()
adegenetTutorial("dapc")
}
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

