% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.snapclust.R
\name{AICc}
\alias{AICc}
\alias{AICc.snapclust}
\title{Compute Akaike Information Criterion for small samples (AICc) for snapclust}
\usage{
AICc(object, ...)

\method{AICc}{snapclust}(object, ...)
}
\arguments{
\item{object}{An object returned by the function \code{\link{snapclust}}.}

\item{...}{Further arguments for compatibility with the \code{AIC} generic
(currently not used).}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\seealso{
\code{\link{snapclust}} to generate clustering solutions.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
