% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.snapclust.R
\name{BIC.snapclust}
\alias{BIC.snapclust}
\title{Compute Bayesian Information Criterion (BIC) for snapclust}
\usage{
\method{BIC}{snapclust}(object, ...)
}
\arguments{
\item{object}{An object returned by the function \code{\link{snapclust}}.}

\item{...}{Further arguments for compatibility with the \code{BIC} generic
(currently not used).}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\seealso{
\code{\link{snapclust}} to generate clustering solutions.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
