% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genclust.em.choose.k.R
\name{genclust.em.choose.k}
\alias{genclust.em.choose.k}
\title{Choose the number of clusters for genclust.em using BIC}
\usage{
genclust.em.choose.k(max, ..., IC = AIC, IC.only = TRUE)
}
\arguments{
\item{...}{Arguments passed to \code{\link{genclust.em}}.}

\item{k.values}{A vector of integers representing the different number of
clusters to try.}

\item{bic.only}{A logical indicating if BIC values only should be returned;
if \code{FALSE}, full \code{genclust.em} objects are returned.}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\code{\link{genclust.em}} to generate individual clustering
    solutions, and \code{\link{BIC.genclust.em}} for computing BIC for
    \code{genclust.em} objects.
}

