% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newclust.R
\name{genclust.em}
\alias{genclust.em}
\title{Maximum-likelihood genetic clustering using EM algorithm}
\usage{
genclust.em(x, k, pop.ini = NULL, max.iter = 100, n.start = 10,
  detailed = TRUE)
}
\arguments{
\item{x}{a \linkS4class{genind} object}

\item{k}{the number of clusters to look for}

\item{pop.ini}{an optional factor defining the initial cluster configuration}

\item{max.iter}{the maximum number of iteration of the EM algorithm}

\item{n.start}{the number of times the EM algorithm is run, each time with different random
starting conditions}

\item{detailed}{a logical stating whether extra details should be incorporated into the output;
these include group membership probability, indication of convergence, and the number of
iterations used before convergence}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\examples{
data(sim2pop)
x <- sim2pop
pop(x) <- sample(pop(x))

## try function using true clusters as initial state
 genclust.em(x, 2, pop.ini = pop(sim2pop), detailed = FALSE)
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com} and Marie-Pauline Beugin
}

