% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genclust.em.R
\name{genclust.em}
\alias{genclust.em}
\title{Maximum-likelihood genetic clustering using EM algorithm}
\usage{
genclust.em(x, k, pop.ini = "kmeans", max.iter = 100, n.start = 10,
  hybrids = FALSE, detailed = TRUE, dim.ini = 100, ...)
}
\arguments{
\item{x}{a \linkS4class{genind} object}

\item{k}{the number of clusters to look for}

\item{pop.ini}{parameter indicating how the initial group membership should
be found. If \code{NULL}, groups are chosen at random, and the algorithm will
be run \code{n.start times}. If "kmeans", then the function
\code{find.clusters} is used to define initial groups. Alternatively, a
factor defining the initial cluster configuration can be provided.}

\item{max.iter}{the maximum number of iteration of the EM algorithm}

\item{n.start}{the number of times the EM algorithm is run, each time with
different random starting conditions}

\item{hybrids}{a logical indicating if hybrids should be modelled
explicitely; this is currently implemented for 2 groups only.}

\item{detailed}{a logical stating whether extra details should be
incorporated into the output; these include group membership probability,
indication of convergence, and the number of iterations used before
convergence}

\item{dim.ini}{the number of PCA axes to retain in the dimension reduction
step for \code{\link{find.clusters}}, if this method is used to define
initial group memberships (see argument \code{pop.ini}).}

\item{...}{further arguments passed on to \code{\link{find.clusters}}}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\examples{
\dontrun{
data(microbov)

## try function using k-means initialization
grp.ini <- find.clusters(microbov, n.clust=15, n.pca=150)

## run EM algo
res <- genclust.em(microbov, 15, pop.ini = grp.ini$grp, detailed = TRUE)
names(res)
res$converged
res$n.iter

## plot result
compoplot(res)

## flag potential hybrids
to.flag <- apply(res$proba,1,max)<.9
compoplot(res, subset=to.flag, show.lab=TRUE,
                 posi="bottomleft", bg="white")


## Simulate hybrids
zebu <- microbov[pop="Zebu"]
salers <- microbov[pop="Salers"]
hyb <- hybridize(zebu, salers, n=30)
x <- repool(zebu, salers, hyb)

## method without hybrids
res.no.hyb <- genclust.em(x, k=2, hybrids=FALSE)
compoplot(res.no.hyb, col.pal=spectral, n.col=2)

## method with hybrids
res.hyb <- genclust.em(x, k=2, hybrids=TRUE)
compoplot(res.hyb, col.pal=spectral, n.col=2)
}
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com} and Marie-Pauline
Beugin
}

