% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KIC.snapclust.R
\name{KIC}
\alias{KIC}
\alias{KIC.snapclust}
\title{Compute Akaike Information Criterion for small samples (AICc) for snapclust}
\usage{
KIC(object, ...)

\method{KIC}{snapclust}(object, ...)
}
\arguments{
\item{object}{An object returned by the function \code{\link{snapclust}}.}

\item{...}{Further arguments for compatibility with the \code{AIC} generic
(currently not used).}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\seealso{
\code{\link{snapclust}} to generate clustering solutions.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
