% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newclust.R
\name{newclust}
\alias{newclust}
\title{Maximum-likelihood genetic clustering}
\usage{
newclust(x, k, pop.ini = NULL, max.iter = 100)
}
\arguments{
\item{x}{a \linkS4class{genind} object}

\item{k}{the number of clusters to look for}

\item{pop.ini}{an optional factor defining the initial cluster configuration}

\item{max.iter}{the maximum number of iteration of the EM algorithm}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\examples{
data(sim2pop)
x <- sim2pop
pop(x) <- sample(pop(x))

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com} and Marie-Pauline Beugin
}

