% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hierarchyMethods.R
\docType{methods}
\name{setpop}
\alias{setpop}
\alias{setpop,genind-method}
\alias{setpop<-}
\alias{setpop<-,genind-method}
\title{Manipulate the population factor of genind objects.}
\usage{
setpop(x, formula = NULL)

setpop(x) <- value
}
\arguments{
\item{x}{a genind object}

\item{formula}{a nested formula indicating the order of the population
hierarchy.}

\item{value}{same as formula}
}
\description{
The following methods allow the user to quickly change the population of a
genind object.
}
\examples{
data(microbov)

sethierarchy(microbov) <- data.frame(other(microbov))

# Currently set on just
head(pop(microbov))

# setting the hierarchy to both Pop and Subpop
setpop(microbov) <- ~coun/breed
head(pop(microbov))

\dontrun{

# Can be used to create objects as well.
microbov.old <- setpop(microbov, ~spe)
head(pop(microbov.old))
}
}
\author{
Zhian N. Kamvar
}

