\name{seqTrack}
\alias{seqTrack}
% \alias{seqTrackG}
\alias{seqTrack.default}
% \alias{seqTrackG.default}
\alias{plotSeqTrack}
% \alias{optimize.seqTrack}
% \alias{optimize.seqTrack.default}
\alias{get.likelihood}
\alias{get.likelihood.seqTrack}
\alias{get.likelihood.seqTrack.default}
\alias{seqTrack-class}
% \alias{get.result.by}
% \alias{get.consensus}
\alias{as,seqTrack,graphNEL-method}
\alias{coerce,seqTrack,graphNEL-method}
\title{SeqTrack algorithm}
\description{
  Important: this method is currently under review. It will be
  documented once accepted for publication.
  Please email the author if you are interested in using it.

}
% \usage{
% %seqTrack(seq.names, seq.dates, W, optim = c("min", "max"), \dots)
% %plotSeqTrack(x, xy, useArrows = TRUE, col = NULL, bg = "grey", add = FALSE, 
%  %   \dots)
% }
% \arguments{
%   % \item{seq.names}{}
%   % \item{seq.dates}{}
%   % \item{W}{}
%   % \item{optim}{}
%   % \item{x}{}
%   % \item{xy}{}
%   % \item{useArrows}{}
%   % \item{col}{}
%   % \item{bg}{}
%   % \item{add}{}
%   % \item{\dots}{}
%   % \item{}{}
% }
\author{Thibaut Jombart \email{t.jombart@imperial.ac.uk}}
