% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapclust.choose.k.R
\name{snapclust.choose.k}
\alias{snapclust.choose.k}
\title{Choose the number of clusters for snapclust using BIC}
\usage{
snapclust.choose.k(max, ..., IC = AIC, IC.only = TRUE)
}
\arguments{
\item{max}{An integer indicating the maximum number of clusters to seek;
\code{\link{snapclust}} will be run for all k from 2 to max.}

\item{...}{Arguments passed to \code{\link{snapclust}}.}

\item{IC}{A function computing the information criterion for
\code{\link{snapclust}} objects. Available statistics are
\code{AIC} (default), \code{AICc}, and \code{BIC}.}

\item{IC.only}{A logical (TRUE by default) indicating if IC values only
should be returned; if \code{FALSE}, full \code{snapclust} objects are
returned.}
}
\description{
Do not use. We work on that stuff. Contact us if interested.
}
\seealso{
\code{\link{snapclust}} to generate individual clustering solutions,
and \code{\link{BIC.snapclust}} for computing BIC for \code{snapclust}
objects.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
