% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/strataMethods.R
\docType{methods}
\name{strata}
\alias{addstrata}
\alias{addstrata,genind-method}
\alias{addstrata,genlight-method}
\alias{addstrata<-}
\alias{addstrata<-,genind-method}
\alias{addstrata<-,genlight-method}
\alias{namestrata}
\alias{namestrata,genind-method}
\alias{namestrata,genlight-method}
\alias{namestrata<-}
\alias{namestrata<-,genind-method}
\alias{namestrata<-,genlight-method}
\alias{splitstrata}
\alias{splitstrata,genind-method}
\alias{splitstrata,genlight-method}
\alias{splitstrata<-}
\alias{splitstrata<-,genind-method}
\alias{splitstrata<-,genlight-method}
\alias{strata}
\alias{strata,genind-method}
\alias{strata,genlight-method}
\alias{strata<-}
\alias{strata<-,genind-method}
\alias{strata<-,genlight-method}
\title{Access and manipulate the population strata for genind or genlight objects.}
\usage{
strata(x, formula = NULL, combine = TRUE, value)

strata(x) <- value

namestrata(x, value)

namestrata(x) <- value

splitstrata(x, value, sep = "_")

splitstrata(x, sep = "_") <- value

addstrata(x, value, name = "NEW")

addstrata(x, name = "NEW") <- value
}
\arguments{
\item{x}{a genind or genlight object}

\item{formula}{a nested formula indicating the order of the population
strata.}

\item{combine}{if \code{TRUE}, the levels will be combined according to the
formula argument. If it is \code{FALSE}, the levels will not be combined.}

\item{value}{a data frame OR vector OR formula (see details).}

\item{sep}{a \code{character} indicating the character used to separate
hierarchical levels. This defaults to "_".}

\item{name}{an optional name argument for use with addstrata if supplying
a vector. Defaults to "NEW".}
}
\description{
The following methods allow the user to quickly change the strata or
population of a genind or genlight object.
}
\details{
\subsection{Function Specifics}{ \itemize{ \item
\strong{strata()} - This will retrieve the data from the
\emph{strata} slot in the \linkS4class{genind} object. You have the
option to choose specific heirarchical levels using a formula (see below) and
you can choose to combine the hierarchical levels (default) \item
\strong{strata()} - Set or reset the hierarchical levels in your
\linkS4class{genind} object. \item \strong{namestrata()} - Rename the
hierarchical levels. \item \strong{splitstrata()} - It is often
difficult to import files with several levels of strata as most data
formats do not allow unlimited population levels. This is circumvented by
collapsing all hierarchical levels into a single population factor with a
common separator for each observation. This function will then split those
hierarchies for you, but it works best on a strata that only has a single
column in it. See the rootrot example below. \item \strong{addstrata()} -
Add levels to your population strata. If you have extra hierarchical
levels you want to add to your population strata, you can use this method
to do so. You can input a data frame or a vector, but if you put in a vector,
you have the option to name it. }}

\subsection{Argument Specifics}{

These functions allow the user to seamlessly assign the hierarchical levels
of their \code{\linkS4class{genind}} object. Note that there are two ways
of performing all methods (except for \code{strata()}). They
essentially do the same thing except that the assignment method (the one with
the "\code{<-}") will modify the object in place whereas the non-assignment
method will not modify the original object. Due to convention, everything
right of the assignment is termed \code{value}. To avoid confusion, here is a
guide to the inputs: \itemize{ \item \strong{strata()} This will be a
\code{\link{data.frame}} that defines the strata for each individual in
the rows. \item \strong{namestrata()} This will be either a
\code{\link{vector}} or a \code{\link{formula}} that will define the names.
\item \strong{splitstrata()} This will be a \code{\link{formula}} argument
with the same number of levels as the strata you wish to split. \item
\strong{addstrata()} This will be a \code{\link{vector}} or
\code{\link{data.frame}} with the same length as the number of individuals in
your data. }}

\subsection{Details on Formulas}{

The preferred use of these functions is with a \code{\link{formula}} object.
Specifically, a hierarchical formula argument is used to assign the levels of
the strata. An example of a hierarchical formula would be:\cr
\code{~Country/City/Neighborhood}\cr or \cr \code{~Country + Country:City +
Country:City:Neighborhood}\cr of course, the first method is slightly easier
to read. It is important to use hiearchical formulas when specifying
hierarchies as other types of formulas (eg.
\code{~Country*City*Neighborhood}) might give spurious results.}
}
\examples{
# let's look at the microbov data set:
data(microbov)
microbov

# We see that we have three vectors of different names in the 'other' slot.
?microbov
# These are Country, Breed, and Species
names(other(microbov))

# Let's set the strata
strata(microbov) <- data.frame(other(microbov))
microbov

# And change the names so we know what they are
namestrata(microbov) <- ~Country/Breed/Species

# let's see what the strata looks like by Species and Breed:
head(strata(microbov, ~Breed/Species))
}
\author{
Zhian N. Kamvar
}
\seealso{
\code{\link{setpop}} \code{\link{genind}}
  \code{\link{as.genind}}
}

