\name{virtualClasses}
\docType{class}
\alias{indInfo-class}
\alias{popInfo-class}
\alias{gen-class}
\alias{callOrNULL-class}
\alias{charOrNULL-class}
\alias{factorOrNULL-class}
\alias{intOrNum-class}
\alias{listOrNULL-class}
\alias{intOrNULL-class}
\title{Virtual classes for adegenet}
\description{These virtual classes are only for internal use in adegenet}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\author{Thibaut Jombart \email{t.jombart@imperial.ac.uk} }
\keyword{classes}
